Budget Controller accepts Budget Items.
Initial commit.

It is possible to add budget items to the budget controller while using the web application. The budget items are not displayed in ther user interface, but are stored in the budget controller. The data stored in the budget controller is logged to the console after each insert. I have focused on the logging part, thus, is should be easy to read the console output from the application.